Moved overlays to correct package
