+++ core/src/com/mygdx/chalmersdefense/views/overlays/SettingsOverlay.java
@@ -1,4 +1,4 @@
-package com.mygdx.chalmersdefense.views.GameScreenViews;
+package com.mygdx.chalmersdefense.views.overlays;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Texture;
@@ -8,8 +8,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.utils.Align;
 import com.mygdx.chalmersdefense.controllers.overlays.SettingsOverlayController;
 import com.mygdx.chalmersdefense.utilities.FontFactory;
-
-import javax.swing.*;
+import com.mygdx.chalmersdefense.views.overlays.AbstractOverlay;
 
 public class SettingsOverlay extends AbstractOverlay {
     private final SettingsOverlayController settingsOverlayController;
@@ -22,7 +21,7 @@ public class SettingsOverlay extends AbstractOverlay {
     }
 
     @Override
-    void initialize() {
+    protected void initialize() {
         stage.addActor(settingsMenuGroup);
         if (!settingsMenuGroup.hasChildren()) {
             settingsMenuGroup.addActor(backgroundImage);
